style: Separate entities by folders
