+++ src/main/java/com/petshopadmin/adapter/input/http/species/SpecieRequestHTTP.java
@@ -1,0 +1,2 @@
+package com.petshopadmin.adapter.input.http.species;public class SpecieRequestHTTP {
+}
